Check for multiple RRSIG entries
So far, only one RRSIG was evaluated if there were multiple - possibly
the wrong one could have been taken which led to false-positive alerts.

This change checks for the availability of multiple RRSIGS on the SOA
and then loops over each of them to check them individually. As soon
as at least one valid RRSIG with a high-enough remaining lifetime is
found, we're happy. (having a second or third RRSIG that is over-
lapping but short before it's end of validity does not hurt - as long
as there is at least one valid RRSIG)